Improve typespec to allow integers to be encoded in RLP
The Yellow Paper speaks about only binaries and lists as the types of data which can be RLP encoded, but the actual test cases from Ethereum allow for integers to be encoded directly (which, I guess are basically the same thing as bitstrings). In either case, we allow for this now in our typespecs.